third_party/blink/web_tests/platform/linux-chrome/external/wpt/pointerevents/pointerevent_capture_suppressing_mouse-expected.txt

This is a testharness.js-based test.
[FAIL] Test pointer capture.
  assert_equals: Move to button from origin: Event  mismatch expected "pointerover@captureButton" but got "pointerout@target1"
Harness: the test ran to completion.
